feat(user): add order invoice and details
- Created invoice view for orders
- Added order details view
- Implemented order search and filtering
- Improved order display in user dashboard
- Updated user order routes

+++ database/migrations/2025_04_24_093709_add_new_fields_to_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('phone')->nullable()->after('country');
             $table->text('address')->nullable()->after('phone');
             $table->string('image')->nullable()->after('address');
-            $table->double('balance', 10, 2)->default(0)->after('image');
+            $table->decimal('balance', 10, 2)->default(0)->after('image');
             $table->enum('status', ['active', 'pending', 'disabled'])->default('pending')->after('balance');
             $table->boolean('email_verify')->default(false)->after('email_verified_at');
             $table->boolean('update_notify')->default(false)->after('email_verify');
